+++ TcManagerWebApp/src/main/java/com/davidlowe/tcmanagerwebapp/services/DocumentComplianceStatusService.java
@@ -4,6 +4,6 @@ package com.davidlowe.tcmanagerwebapp.services;
 import com.davidlowe.tcmanagerwebapp.models.DocumentComplianceStatus;
 
 
-public interface DocumentComplianceStatusService extends ReadOnlyService<DocumentComplianceStatus, Long>
+public interface DocumentComplianceStatusService extends ReadOnlyService<DocumentComplianceStatus, Integer>
 {
 }
