+++ TcManagerWebApp/src/main/java/com/davidlowe/tcmanagerwebapp/services/ListingService.java
@@ -4,6 +4,6 @@ package com.davidlowe.tcmanagerwebapp.services;
 import com.davidlowe.tcmanagerwebapp.models.Listing;
 
 
-public interface ListingService extends CrudService<Listing, Long>
+public interface ListingService extends CrudService<Listing, Integer>
 {
 }
