+++ TcManagerWebApp/src/main/java/com/davidlowe/tcmanagerwebapp/services/ListingStatusService.java
@@ -4,6 +4,6 @@ package com.davidlowe.tcmanagerwebapp.services;
 import com.davidlowe.tcmanagerwebapp.models.ListingStatus;
 
 
-public interface ListingStatusService extends ReadOnlyService<ListingStatus, Long>
+public interface ListingStatusService extends ReadOnlyService<ListingStatus, Integer>
 {
 }
