+++ TcManagerWebApp/src/main/java/com/davidlowe/tcmanagerwebapp/services/PersonService.java
@@ -4,6 +4,6 @@ package com.davidlowe.tcmanagerwebapp.services;
 import com.davidlowe.tcmanagerwebapp.models.Person;
 
 
-public interface PersonService extends CrudService<Person, Long>
+public interface PersonService extends CrudService<Person, Integer>
 {
 }
